refactor: apply auto-format and pass tests
