+++ game/core/game_view.py
@@ -11,16 +11,15 @@ from beartype import beartype
 from loguru import logger
 from pyglet.math import Vec2
 
-from game.core.pause_menu import PauseMenu
-
 from .constants import CAMERA_SPEED, HORIZONTAL_MARGIN, MAP_SIZE, VERTICAL_MARGIN
 from .game_clock import GameClock
-from .gui import GameGUI
 from .game_map import GameMap
 from .game_state import GameState
+from .gui import GameGUI
+from .gui.rpg_movement import RPGMovement
+from .pause_menu import PauseMenu
 from .pressed_keys import PressedKeys
 from .registration import Register, SpriteRegister
-from .view_strategies.rpg_movement import RPGMovement
 
 
 class GameView(arcade.View):
