+++ game/core/gui/inventory.py
@@ -1,4 +1,4 @@
-"""Inventory"""
+"""Inventory."""
 
 import arcade
