+++ game/core/view_strategies/raft_movement.py
@@ -17,15 +17,13 @@ def check_missing_components(inventory):
 
 
 def generate_missing_components_text(inventory):
-    expected = []
+    expected = [f"{count} {component}" for component, count in RAFT_COMPONENTS.items()]
     actual = []
-    conj = " and "
-    for component, count in RAFT_COMPONENTS.items():
-        expected.append(f"{count} {component}")
     for item in inventory:
         name = item.properties["name"]
         if name in RAFT_COMPONENTS:
             actual.append(f"{item.properties['count']} {name}")
+    conj = " and "
     return {
         "message": "You attemped to build a raft and was unsuccessful",
         "notes": f"Raft requires {conj.join(expected)}. You have {conj.join(actual)}.",
