+++ game/tasks/__init__.py
@@ -1,6 +1,6 @@
 """Internal logic for pointing to whichever task code will be loaded."""
 
-from . import task01_s_player as player_module  # noqa: F401
+from . import task01_s_select_player as player_module  # noqa: F401
 from . import (
     task02_o_inventory,
     task03_l_crafting,
