+++ game/tasks/task03_l_crafting.py
@@ -26,6 +26,7 @@ SOURCE_NAME = "task03_L_TBD"  # FYI: Required for code reload
 def load_sprites(sprite_register: SpriteRegister) -> None:
     """Create the special 'player sprite' who can be moved with WASD or the arrow keys."""
 
+
 """
 
 If you would like to learn more about the Liskov-Substitution Principle, you can take a look at:
