+++ game/tasks/task04_i_the_raft.py
@@ -23,6 +23,7 @@ SOURCE_NAME = "task04_I_TBD"  # FYI: Required for code reload
 def load_sprites(sprite_register: SpriteRegister) -> None:
     """Create the special 'player sprite' who can be moved with WASD or the arrow keys."""
 
+
 """
 
 If you would like to learn more about the Interface Segregation Principle, you can take a look at:
