+++ game/tasks/task05_d_the_new_island.py
@@ -23,6 +23,7 @@ SOURCE_NAME = "task05_D_TBD"  # FYI: Required for code reload
 def load_sprites(sprite_register: SpriteRegister) -> None:
     """Create the special 'player sprite' who can be moved with WASD or the arrow keys."""
 
+
 """
 
 If you would like to learn more about the Dependency Inversion Principle, you can take a look at:
