+++ requirements.txt
@@ -78,11 +78,11 @@ nodeenv==1.7.0 ; python_full_version >= "3.10.5" and python_full_version < "4.0.
 numpy==1.23.4 ; python_full_version >= "3.10.5" and python_full_version < "4.0.0"
 packaging==21.3 ; python_full_version >= "3.10.5" and python_full_version < "4.0.0"
 pandas==1.5.1 ; python_full_version >= "3.10.5" and python_full_version < "4.0.0"
-pathspec==0.10.1 ; python_full_version >= "3.10.5" and python_full_version < "4.0.0"
-pattern-feedback-tool==0.3.10 ; python_full_version >= "3.10.5" and python_full_version < "4.0.0"
+pathspec==0.10.2 ; python_full_version >= "3.10.5" and python_full_version < "4.0.0"
+pattern-feedback-tool==0.3.12 ; python_full_version >= "3.10.5" and python_full_version < "4.0.0"
 pbr==5.11.0 ; python_full_version >= "3.10.5" and python_full_version < "4.0.0"
 pillow==9.1.1 ; python_full_version >= "3.10.5" and python_full_version < "4.0.0"
-platformdirs==2.5.3 ; python_full_version >= "3.10.5" and python_full_version < "4.0.0"
+platformdirs==2.5.4 ; python_full_version >= "3.10.5" and python_full_version < "4.0.0"
 pluggy==1.0.0 ; python_full_version >= "3.10.5" and python_full_version < "4.0.0"
 pre-commit==2.20.0 ; python_full_version >= "3.10.5" and python_full_version < "4.0.0"
 pycg==0.0.6 ; python_full_version >= "3.10.5" and python_full_version < "4.0.0"
@@ -127,7 +127,7 @@ typing-inspect==0.8.0 ; python_full_version >= "3.10.5" and python_full_version 
 unimport==0.12.2 ; python_full_version >= "3.10.5" and python_full_version < "4.0.0"
 untokenize==0.1.1 ; python_full_version >= "3.10.5" and python_version < "4.0"
 urllib3==1.26.12 ; python_full_version >= "3.10.5" and python_version < "4"
-virtualenv==20.16.6 ; python_full_version >= "3.10.5" and python_full_version < "4.0.0"
+virtualenv==20.16.7 ; python_full_version >= "3.10.5" and python_full_version < "4.0.0"
 vulture==2.6 ; python_full_version >= "3.10.5" and python_full_version < "4.0.0"
 watchdog==2.1.9 ; python_full_version >= "3.10.5" and python_full_version < "4.0.0"
 win32-setctime==1.1.0 ; python_full_version >= "3.10.5" and python_full_version < "4.0.0" and sys_platform == "win32"
